repair 0 new

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -2,7 +2,7 @@
 
 void print_binary(unsigned long int n)
 {
-	unsigned int mask = 32768;
+	unsigned long int mask = 32768;
 
 	while(mask > 1)
 	{
